+++ courses/database/migrations/2019_08_19_000000_create_failed_jobs_table.php
@@ -9,17 +9,20 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
+//    protected $connection = 'tenant';
     public function up(): void
     {
-        Schema::create('failed_jobs', function (Blueprint $table) {
-            $table->id();
-            $table->string('uuid')->unique();
-            $table->text('connection');
-            $table->text('queue');
-            $table->longText('payload');
-            $table->longText('exception');
-            $table->timestamp('failed_at')->useCurrent();
-        });
+        if (Schema::getConnection()->getDatabaseName() === env("DB_DATABASE")) {
+            Schema::connection('tenant')->create('failed_jobs', function (Blueprint $table) {
+                $table->id();
+                $table->string('uuid')->unique();
+                $table->text('connection');
+                $table->text('queue');
+                $table->longText('payload');
+                $table->longText('exception');
+                $table->timestamp('failed_at')->useCurrent();
+            });
+        }
     }
 
     /**
@@ -27,6 +30,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('failed_jobs');
+        if (Schema::getConnection()->getDatabaseName() === env("DB_DATABASE")) {
+            Schema::dropIfExists('failed_jobs');
+        }
     }
 };
